Move Robolectric-only resource tests back to ShadowResourcesTest
ctesque ResourcesTest cannot depend on Robolectric.buildAttributeSet,
which is only available in a Robolectric environment. ctesque
ResourcesTest is designed to work across all execution environments
(e.g. emulators).
